# docs/openapi.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/v1/viagens/": {
      "post": {
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ViagemResponse"
            }
          }
        },
        "operationId": "create_viagem",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ViagemCreateRequest"
            }
          }
        ],
        "tags": [
          "viagens"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ViagemListResponse"
            }
          }
        },
        "operationId": "list_all_viagens",
        "parameters": [
          {
            "name": "data_inicio",
            "in": "query",
            "type": "string",
            "description": "Data inicial (YYYY-MM-DD)"
          },
          {
            "name": "data_fim",
            "in": "query",
            "type": "string",
            "description": "Data final (YYYY-MM-DD)"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "description": "Status da viagem"
          },
          {
            "name": "motorista_id",
            "in": "query",
            "type": "string",
            "description": "UUID do motorista"
          },
          {
            "name": "rota_id",
            "in": "query",
            "type": "string",
            "description": "UUID da rota"
          }
        ],
        "tags": [
          "viagens"
        ]
      }
    },
...
    "/v1/viagens/gerar-lote": {
      "post": {
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ViagemLoteResponse"
            }
          }
        },
        "operationId": "gerar_viagens_lote",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ViagemLoteRequest"
            }
          }
        ],
        "tags": [
          "viagens"
        ]
      }
    },
...
    "/v1/viagens/{id}/acao": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ViagemResponse"
            }
          }
        },
        "operationId": "control_viagem",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ViagemAcaoRequest"
            }
          }
        ],
        "tags": [
          "viagens"
        ]
      }
    },
    "/v1/viagens/{id}/cancelar": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Trip cancelled successfully"
          }
        },
        "summary": "Cancela uma viagem agendada e notifica alunos (Gestor)",
        "operationId": "cancelar_viagem",
        "tags": [
          "viagens"
        ]
      }
    },
    "/v1/viagens/{id}/confirmacao": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ViagemAlunoConfirmacaoResponse"
            }
          }
        },
        "operationId": "confirmar_presenca",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ViagemConfirmacaoRequest"
            }
          }
        ],
        "tags": [
          "viagens"
        ]
      }
    },
    "/v1/viagens/{id}/localizacao": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "(Motorista) Envia coordenada GPS atual do ônibus em tempo real",
        "operationId": "atualizar_localizacao_onibus",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LocalizacaoRequest"
            }
          }
        ],
        "tags": [
          "viagens"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "(Aluno/Motorista) Obtém a localização atual do ônibus (viagem em andamento)",
        "operationId": "obter_localizacao_onibus",
        "tags": [
          "viagens"
        ]
      }
    },
    "/v1/viagens/{id}/pontos-embarque": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PontoFlatListResponse"
            }
          }
        },
        "summary": "Lista todos os pontos de embarque disponíveis para esta viagem",
        "operationId": "list_pontos_embarque_viagem",
        "tags": [
          "viagens"
        ]
      }
    },
    "/v1/viagens/{viagem_id}/localizacao-aluno": {
      "parameters": [
        {
          "name": "viagem_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Recebe o GPS em tempo real do Aluno para o Auto-Checkin (Geofencing)",
        "operationId": "atualizar_localizacao_aluno",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LocalizacaoRequest"
            }
          }
        ],
        "security": [
          {
            "Bearer Auth": []
          }
        ],
        "tags": [
          "viagens"
        ]
      }
    }
  },
  "info": {
    "title": "BusKá API",
    "version": "1.0.0",
    "description": "\n## Sistema de Gerenciamento de Transporte Escolar\n\nAPI para gerenciamento de rotas, viagens e alunos do transporte escolar municipal.\n\n### Autenticação\nTodos os endpoints (exceto `/auth/login`) requerem autenticação JWT.\nInclua o header: `Authorization: Bearer <seu_token>`\n\n### Roles\n- **ALUNO**: Visualiza rotas, confirma presença em viagens\n- **MOTORISTA**: Gerencia viagens atribuídas, inicia/finaliza trajetos\n- **GESTOR**: Acesso completo à prefeitura (CRUD de rotas, motoristas, relatórios)\n        "
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: Bearer <token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "definitions": {
...
    "PontoFlatListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PontoFlatResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de pontos"
        }
      },
      "type": "object"
    },
    "PontoFlatResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID do ponto"
        },
        "apelido": {
          "type": "string",
          "description": "Nome/apelido"
        },
        "latitude": {
          "type": "number",
          "description": "Latitude"
        },
        "longitude": {
          "type": "number",
          "description": "Longitude"
        },
        "ordem": {
          "type": "integer",
          "description": "Ordem"
        }
      },
      "type": "object"
    },
...
    "ViagemConfirmacaoRequest": {
      "required": [
        "confirmacao"
      ],
      "properties": {
        "confirmacao": {
          "type": "boolean",
          "description": "Confirmado"
        },
        "ponto_embarque_id": {
          "type": "string",
          "description": "UUID do ponto de embarque (opcional)"
        }
      },
      "type": "object"
    },
    "ViagemAlunoConfirmacaoResponse": {
      "properties": {
        "aluno_id": {
          "type": "string",
          "description": "UUID do aluno"
        },
        "nome": {
          "type": "string",
          "description": "Nome do aluno"
        },
        "confirmacao": {
          "type": "boolean",
          "description": "Confirmado"
        },
        "ponto_embarque": {
          "type": "string",
          "description": "UUID do ponto de embarque"
        }
      },
      "type": "object"
    },
    "ViagemCreateRequest": {
      "required": [
        "data",
        "horario_id",
        "rota_id"
      ],
      "properties": {
        "rota_id": {
          "type": "string",
          "description": "UUID da rota"
        },
        "horario_id": {
          "type": "string",
          "description": "UUID do horário"
        },
        "data": {
          "type": "string",
          "description": "Data (YYYY-MM-DD)"
        },
        "motorista_id": {
          "type": "string",
          "description": "UUID do motorista (opcional)"
        },
        "veiculo_id": {
          "type": "string",
          "description": "UUID do veículo (opcional)"
        }
      },
      "type": "object"
    },
    "ViagemResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID"
        },
        "data": {
          "type": "string",
          "description": "Data"
        },
        "horario_saida": {
          "type": "string",
          "description": "Horário"
        },
        "sentido": {
          "type": "string",
          "description": "Sentido"
        },
        "status": {
          "type": "string",
          "description": "AGENDADA, EM_ANDAMENTO, FINALIZADA"
        },
        "rota_id": {
          "type": "string",
          "description": "UUID da rota"
        },
        "rota_nome": {
          "type": "string",
          "description": "Nome da rota"
        },
        "motorista_id": {
          "type": "string",
          "description": "UUID do motorista"
        },
        "veiculo_id": {
          "type": "string",
          "description": "UUID do veículo"
        }
      },
      "type": "object"
    },
    "ViagemListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ViagemResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de viagens"
        }
      },
      "type": "object"
    },
    "ViagemLoteRequest": {
      "required": [
        "data"
      ],
      "properties": {
        "data": {
          "type": "string",
          "description": "Data para gerar viagens (YYYY-MM-DD)"
        }
      },
      "type": "object"
    },
    "ViagemLoteResponse": {
      "properties": {
        "total_rotas_analisadas": {
          "type": "integer",
          "description": "Total de rotas analisadas"
        },
        "viagens_criadas": {
          "type": "integer",
          "description": "Total de viagens criadas"
        },
        "detalhes": {
          "type": "array",
          "items": {
            "type": "string",
            "description": "Detalhes da viagem"
          }
        }
      },
      "type": "object"
    },
    "ViagemAcaoRequest": {
      "required": [
        "acao"
      ],
      "properties": {
        "acao": {
          "type": "string",
          "description": "iniciar ou finalizar"
        }
      },
      "type": "object"
    },
    "LocalizacaoRequest": {
      "required": [
        "latitude",
        "longitude"
      ],
      "properties": {
        "latitude": {
          "type": "number",
          "description": "Latitude atual do ônibus"
        },
        "longitude": {
          "type": "number",
          "description": "Longitude atual do ônibus"
        }
      },
      "type": "object"
    },
...
  },
...
}